refactor(http/upgrade): internal interfaces are private (#3530)
the `linkerd-http-upgrade` crate supports HTTP/1.1 upgrades. currently, it
exposes a number of functions and types in its public interface that are
strictly internal helpers. this branch redefines various interfaces and fields
as private, to clarify the public interface of the crate.

---

* refactor(proxy/http): `h1::is_upgrade()` is private

this function is not used elsewhere, so it does not have to be
`pub(crate)`.

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(http/upgrade): `halves()` is private

this commit restricts `Http11Upgrade::halves` so that it is now a
private interface.

this is an internal interface that is not used elsewhere.

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(http/upgrade): `Http11UpgradeHalves` is private

with `halves()` having been made private, this type is also not used by
other external code. we can make it, and its constituent fields,
private.

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(http/upgrade): `UpgradeBody::upgrade` is private

this commit makes the `upgrade` field of the `UpgradeBody` type private.
it contains two pieces of state that are used in the `Drop`
implementation of an upgrade body.

because these are not accessed or modified elsewhere, this field can be
made private.

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(http/upgrade): `UpgradeBody::new` is `pub(crate)`

this function is not, and *should* not, be used by external callers. we
install the `UpgradeBody` in `Service`.

Signed-off-by: katelyn martin <kate@buoyant.io>

---------

Signed-off-by: katelyn martin <kate@buoyant.io>